Move selectors and update imports
