+++ ui/src/AtlasUI/store/index.js
@@ -1,2 +1,5 @@
+import * as selectors from './selectors';
+
 export { activitySearchActions, activitySearchReducer } from './activitySearch';
 export { activityActions, activityReducer } from './activity';
+export { selectors };
