+++ ui/src/components/editor/Tools/ClearTool/connect.js
@@ -1,9 +1,10 @@
 import { connect } from 'react-redux';
 
 import { ClearTool } from './ClearTool';
-import { selectEditorClearMode } from '../../../../store/selectors';
+import { selectors } from 'AtlasUI/store';
 import { actions } from '../../../../implementations/kapture-backend/actions/';
 
+const { selectEditorClearMode } = selectors;
 const { setClearMode } = actions.editorTools;
 
 const mapState = (state, props) => {
