+++ ui/src/implementations/kapture-backend/actions/activityActions.js
@@ -1,6 +1,5 @@
 import bigInt from 'big-integer';
-import { activityActions } from 'AtlasUI/store';
-import { selectors } from '../store';
+import { activityActions, selectors } from 'AtlasUI/store';
 import {
   REQUEST_PENDING,
   REQUEST_SUCCESS,
