+++ ui/src/implementations/kapture-backend/routes/activities/connect.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 
 import { Activities } from './Activities';
-import { selectors } from '../../store';
+import { selectors } from 'AtlasUI/store';
 import { actions } from '../../actions';
 
 const { loadActivity, resetActivity, publishActivityGrids } = actions.activity;
