+++ ui/src/store/index.js
@@ -1,9 +1,7 @@
-import * as selectors from './selectors';
 import * as plateFunctions from './plateFunctions';
 export { store } from './store';
 export { editorActions } from './editor';
 export { editorComponentsActions } from './editorComponents';
 export { editorToolsActions } from './editorTools';
 export { printActions } from './print';
-export { selectors };
 export { plateFunctions };
